pyproject.toml
[project]
name = "order-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.3"
dependencies = [
    "amqp==5.3.1",
    "asgiref==3.10.0",
    "billiard==4.2.2",
    "celery==5.5.3",
    "certifi==2025.10.5",
    "charset-normalizer==3.4.4",
    "click==8.3.0",
    "click-didyoumean==0.3.1",
    "click-plugins==1.1.1.2",
    "click-repl==0.3.0",
    "django==5.2.7",
    "django-cors-headers>=4.9.0",
    "django-coverage-plugin>=3.2.0",
    "django-debug-toolbar>=6.1.0",
    "django-filter>=25.2",
    "django-jet-reboot>=1.3.10",
    "django-rest-passwordreset==1.5.0",
    "djangorestframework==3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "idna==3.11",
    "kombu==5.5.4",
    "packaging==25.0",
    "prompt-toolkit==3.0.52",
    "psycopg2>=2.9.11",
    "python-dateutil==2.9.0.post0",
    "python-dotenv>=1.1.1",
    "pyyaml==6.0.3",
    "redis>=7.0.1",
    "requests==2.32.5",
    "six==1.17.0",
    "social-auth-app-django>=5.6.0",
    "sqlparse==0.5.3",
    "tzdata==2025.2",
    "ujson==5.11.0",
    "urllib3==2.5.0",
    "vine==5.1.0",
    "wcwidth==0.2.14",
]

[tool.setuptools]
packages = ["orders"]

[dependency-groups]
dev = [
    "coverage>=7.12.0",
    "codecov>=2.0",
]

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"
